+++ apps/sim/app/api/table/[tableId]/rows/find/route.test.ts
@@ -23,7 +23,7 @@ vi.mock('@/app/api/table/utils', async () => {
   }
 })
 
-vi.mock('@/lib/table/service', () => ({
+vi.mock('@/lib/table/rows/service', () => ({
   findRowMatches: mockFindRowMatches,
 }))
