+++ apps/sim/app/api/table/[tableId]/rows/route.ts
@@ -25,7 +25,7 @@ import {
   validateRowData,
   validateRowSize,
 } from '@/lib/table'
-import { queryRows } from '@/lib/table/service'
+import { queryRows } from '@/lib/table/rows/service'
 import { TableQueryValidationError } from '@/lib/table/sql'
 import { rowWireTranslators } from '@/app/api/table/row-wire'
 import { accessError, checkAccess, rowWriteErrorResponse } from '@/app/api/table/utils'
