+++ apps/sim/lib/copilot/tools/server/table/user-table.test.ts
@@ -58,24 +58,27 @@ vi.mock('@/enrichments/registry', () => ({
 vi.mock('@/lib/table/service', () => ({
   addTableColumn: vi.fn(),
   addWorkflowGroup: mockAddWorkflowGroup,
-  batchInsertRows: mockBatchInsertRows,
-  batchUpdateRows: vi.fn(),
   createTable: mockCreateTable,
   deleteColumn: vi.fn(),
   deleteColumns: vi.fn(),
+  deleteTable: mockDeleteTable,
+  getTableById: mockGetTableById,
+  renameColumn: vi.fn(),
+  renameTable: vi.fn(),
+  updateColumnConstraints: vi.fn(),
+  updateColumnType: vi.fn(),
+}))
+
+vi.mock('@/lib/table/rows/service', () => ({
+  batchInsertRows: mockBatchInsertRows,
+  batchUpdateRows: vi.fn(),
   deleteRow: vi.fn(),
   deleteRowsByFilter: vi.fn(),
   deleteRowsByIds: vi.fn(),
-  deleteTable: mockDeleteTable,
   getRowById: vi.fn(),
-  getTableById: mockGetTableById,
   insertRow: vi.fn(),
   queryRows: vi.fn(),
-  renameColumn: vi.fn(),
-  renameTable: vi.fn(),
   replaceTableRows: mockReplaceTableRows,
-  updateColumnConstraints: vi.fn(),
-  updateColumnType: vi.fn(),
   updateRow: vi.fn(),
   updateRowsByFilter: vi.fn(),
 }))
