+++ apps/sim/lib/copilot/tools/server/table/user-table.ts
@@ -29,31 +29,33 @@ import {
 } from '@/lib/table/column-keys'
 import { columnTypeForLeaf, deriveOutputColumnName } from '@/lib/table/column-naming'
 import {
-  addTableColumn,
-  addWorkflowGroup,
-  addWorkflowGroupOutput,
   batchInsertRows,
   batchUpdateRows,
-  createTable,
-  deleteColumn,
-  deleteColumns,
   deleteRow,
   deleteRowsByFilter,
   deleteRowsByIds,
+  getRowById,
+  insertRow,
+  queryRows,
+  replaceTableRows,
+  updateRow,
+  updateRowsByFilter,
+} from '@/lib/table/rows/service'
+import {
+  addTableColumn,
+  addWorkflowGroup,
+  addWorkflowGroupOutput,
+  createTable,
+  deleteColumn,
+  deleteColumns,
   deleteTable,
   deleteWorkflowGroup,
   deleteWorkflowGroupOutput,
-  getRowById,
   getTableById,
-  insertRow,
-  queryRows,
   renameColumn,
   renameTable,
-  replaceTableRows,
   updateColumnConstraints,
   updateColumnType,
-  updateRow,
-  updateRowsByFilter,
   updateWorkflowGroup,
 } from '@/lib/table/service'
 import type {
