+++ apps/sim/lib/table/__tests__/update-row.test.ts
@@ -5,13 +5,12 @@ import { dbChainMock, dbChainMockFns, resetDbChainMock } from '@sim/testing'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 import {
   batchInsertRows,
-  deleteColumn,
   insertRow,
-  renameColumn,
   replaceTableRows,
   updateRow,
   upsertRow,
-} from '@/lib/table/service'
+} from '@/lib/table/rows/service'
+import { deleteColumn, renameColumn } from '@/lib/table/service'
 import type { TableDefinition } from '@/lib/table/types'
 import { getUniqueColumns } from '@/lib/table/validation'
