+++ apps/sim/lib/table/index.ts
@@ -11,6 +11,7 @@ export * from './constants'
 export * from './import'
 export * from './llm'
 export * from './query-builder'
+export * from './rows/service'
 export * from './service'
 export * from './sql'
 export * from './types'
